project/pyproject.toml
[tool.poetry]
name = "mis-service"
version = "0.1.0"
description = "Примитивный веб-сервис для Медицинской Информационной Системы"
authors = ["Artem Bazeev <artbazeev@gmail.com>"]
license = "proprietary"

[tool.poetry.dependencies]
python = "^3.12"
celery = "^5.5.3"
django = "^5.1.2"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
django-filter = "^24.3"
django-lifecycle = "^1.1.2"
django-phonenumber-field = "^7.2.0"
django-querycount = "^0.8.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-extra-fields = "^3.7.0"
drf-yasg = "^1.21.7"
flake8 = "^7.1.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
redis = "^5.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"

[tool.poetry.scripts]
test = "pytest core"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
